lesson7/pyproject.toml
[tool.poetry]
name = "fastapi-llm"
version = "0.1.0"
description = "This is an application in Information Retrieval using FastAPI and LLM"
authors = ["phamdinhdat-ai <datolg123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.5"}
fastcrud = "^0.15.1"
sqlmodel = "^0.0.22"
openai = "^1.55.1"
aiosqlite = "^0.20.0"
greenlet = "^3.1.1"
python-jose = "^3.3.0"
bcrypt = "^4.2.1"
torch = "^2.5.1"
datasets = "^3.1.0"
transformers = "^4.46.3"
huggingface = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
